Fix: Address PR review comments

+++ lib/src/utils/calculation_utils.dart
@@ -4,14 +4,12 @@ class CalculationUtils {
   /// Calculates the net balance for each user across all given expenses.
   /// A positive balance means the user is owed money (creditor).
   /// A negative balance means the user owes money (debtor).
-  static Map<String, double> calculateBalances(List<Expense> expenses) {
-    Map<String, double> balances = {};
+  static Map<String, int> calculateBalances(List<Expense> expenses) {
+    Map<String, int> balances = {};
 
     for (var expense in expenses) {
-      // The payer gets credited the total amount they paid.
       balances[expense.payerId] = (balances[expense.payerId] ?? 0) + expense.amount;
 
-      // Each person in the split gets debited their share.
       for (var split in expense.splits) {
         balances[split.userId] = (balances[split.userId] ?? 0) - split.amount;
       }
